chore: remove useless caption (#6250)
That the only table with a caption, the text above is enough to get what
the table is.

eg:

```
Attributes List

Attributes are properties of pull requests or events used to filter and
evaluate the conditions within rules. They can be used to create more
specific and targeted rules for your repository.
```

Depends-On: #6249